refactor(core): 将多处数值类型由 double 替换为 BigDecimal
- 将 CanaryEvaluator 中灰度比例相关比较改为 BigDecimal 比较逻辑
- 将多处 VO 类中的 double 类型字段（如 metric、threshold、triggerRate、confidence 等）改为 BigDecimal 类型
- 统一 RuleDefinition、RuleDefinitionDTO、ScorecardDefinitionDTO 中灰度比例及评分相关字段为 BigDecimal 类型
- 修改服务层、存储库及转换逻辑中对应数值字段使用 BigDecimal，确保数值精度一致
- 调整 ScorecardRule 中的 BigDecimal 转 double 处理，兼容计算引擎双精度需求
- 细化方法注释，改进代码可读性与一致性
- 其他格式优化，如 HTMLFilter 中 switch-case 结构调整，注解文档补充等

+++ ydsz-literule/ydsz-literule-domain/src/main/java/com/njydsz/literule/domain/vo/CEPHitVO.java
@@ -1,5 +1,6 @@
 package com.njydsz.literule.domain.vo;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
@@ -30,7 +31,7 @@ public class CEPHitVO {
   private Instant hitAt;
 
   /** 命中指标值（如窗口内聚合度量，用于排序/告警分级） */
-  private double metric;
+  private BigDecimal metric;
 
   /** 命中上下文（附加维度信息，如项目/组织等） */
   private Map<String, Object> context;
